requirements.txt
Django
djangorestframework
psycopg2>=2.7.5,<2.8.0
django-filter==2.3.0
